Ticket #6
Server: implementare ServiceImpl

+++ rubrica-gruppo3-server/src/main/java/it/iad/rubricagruppo3/service/RubricaService.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 
 public interface RubricaService {
-    int conta();
+    long conta();
     List<Contatto> search(String criterio);
     List<Contatto> cancellaRubrica(Long id);
     List<Contatto> aggiungiRubrica(Contatto contatto);
